pyproject.toml
[tool.poetry]
name = "human-in-the-loop"
version = "0.1.0"
description = ""
authors = ["Yaroslav Kovaliv <tlcees@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
black = "^25.1.0"
isort = "^6.0.0"
langchain = "^0.3.17"
langchain-openai = "^0.3.3"
langgraph = "^0.2.69"
langchainhub = "^0.1.21"
langgraph-checkpoint-sqlite = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
